refactor DRY
