+++ antelope-common/src/lib.rs
@@ -1,3 +1,2 @@
-mod actions;
 mod index;
 mod maps;
